test: 통합 테스트 코드 작성(미완)
